CI: add unit tests
Now that we moved core components to a library and using functions, it
is much easier to perform unit tests to ensure some basic correctness.

This commit adds a new tests directory with tests for all the various
part of the software, it is nowhere near complete, but it's a start.

Added also the action to perform them for each commit.

Signed-off-by: Luca Di Maio <luca.dimaio1@gmail.com>